backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.1
langchain==0.0.335
openai==1.2.4
python-dotenv==1.0.0 